Resolve conflict between two different uses of term 'dataset'.
Now noms/Dataset just refers to the client interface to a particular dataset.

noms/mgmt/Dataset is an internal implementation detail which is a strong type wrapped around the Dataset entries inside the database.
