+++ clients/counter/counter.go
@@ -10,10 +10,10 @@ import (
 )
 
 func main() {
-	datasetDataStoreFlags := dataset.DatasetDataFlags()
+	dsFlags := dataset.Flags()
 	flag.Parse()
 
-	ds := datasetDataStoreFlags.CreateStore()
+	ds := dsFlags.CreateDataset()
 	if ds == nil {
 		flag.Usage()
 		return
