+++ dataset/dataset.go
@@ -1,54 +1,79 @@
 package dataset
 
 import (
+	"flag"
+
+	"github.com/attic-labs/noms/chunks"
 	"github.com/attic-labs/noms/datas"
-	. "github.com/attic-labs/noms/dbg"
+	"github.com/attic-labs/noms/dataset/mgmt"
+	"github.com/attic-labs/noms/ref"
 	"github.com/attic-labs/noms/types"
 )
 
-//go:generate go run gen/types.go -o types.go
+// TODO: Could just literally use datastore and just refer to it as 'dataset'. Unsure which is better.
+type Dataset struct {
+	datas.DataStore
+}
 
-func GetDatasets(ds datas.DataStore) DatasetSet {
-	if ds.Roots().Empty() {
-		return NewDatasetSet()
-	} else {
-		// BUG 13: We don't ever want to branch the datasets database. Currently we can't avoid that, but we should change DataStore::Commit() to support that mode of operation.
-		Chk.EqualValues(1, ds.Roots().Len())
-		return DatasetSetFromVal(ds.Roots().Any().Value())
+func NewDataset(rootStore datas.DataStore, datasetID string) Dataset {
+	return Dataset{datas.NewDataStore(rootStore, &datasetRootTracker{rootStore, datasetID})}
+}
+
+func (ds *Dataset) Commit(newRoots datas.RootSet) Dataset {
+	return Dataset{ds.DataStore.Commit(newRoots)}
+}
+
+type datasetFlags struct {
+	chunks.Flags
+	datasetID *string
+}
+
+func Flags() datasetFlags {
+	return datasetFlags{
+		chunks.NewFlags(),
+		flag.String("dataset-id", "", "dataset id to store data for"),
 	}
 }
 
-func CommitDatasets(ds datas.DataStore, datasets DatasetSet) datas.DataStore {
-	return ds.Commit(datas.NewRootSet().Insert(
-		datas.NewRoot().SetParents(
-			ds.Roots().NomsValue()).SetValue(
-			datasets.NomsValue())))
-}
-
-func getDataset(datasets DatasetSet, datasetID string) (r *Dataset) {
-	datasets.Iter(func(dataset Dataset) (stop bool) {
-		if dataset.Id().String() == datasetID {
-			r = &dataset
-			stop = true
-		}
-		return
-	})
-	return
-}
-
-func GetDatasetRoot(datasets DatasetSet, datasetID string) types.Value {
-	dataset := getDataset(datasets, datasetID)
-	if dataset == nil {
+func (f datasetFlags) CreateDataset() *Dataset {
+	if *f.datasetID == "" {
 		return nil
 	}
-	return dataset.Root()
+	cs := f.Flags.CreateStore()
+	if cs == nil {
+		return nil
+	}
+
+	// Blech, kinda sucks to typecast to RootTracker, but we know that all the implementations of ChunkStore that implement it.
+	rootDataStore := datas.NewDataStore(cs, cs.(chunks.RootTracker))
+
+	ds := NewDataset(rootDataStore, *f.datasetID)
+	return &ds
 }
 
-func SetDatasetRoot(datasets DatasetSet, datasetID string, val types.Value) DatasetSet {
-	newDataset := NewDataset().SetId(types.NewString(datasetID)).SetRoot(val)
-	dataset := getDataset(datasets, datasetID)
+// TODO: Move to separate file
+type datasetRootTracker struct {
+	rootStore datas.DataStore
+	datasetID string
+}
+
+func (rt *datasetRootTracker) Root() ref.Ref {
+	dataset := mgmt.GetDatasetRoot(mgmt.GetDatasets(rt.rootStore), rt.datasetID)
 	if dataset == nil {
-		return datasets.Insert(newDataset)
+		return ref.Ref{}
+	} else {
+		return dataset.Ref()
 	}
-	return datasets.Remove(*dataset).Insert(newDataset)
+}
+
+func (rt *datasetRootTracker) UpdateRoot(current, last ref.Ref) bool {
+	if last != rt.Root() {
+		return false
+	}
+
+	// BUG 11: Sucks to have to read the dataset root here in order to commit.
+	datasetRoot := types.MustReadValue(current, rt.rootStore)
+
+	rt.rootStore = mgmt.CommitDatasets(rt.rootStore, mgmt.SetDatasetRoot(mgmt.GetDatasets(rt.rootStore), rt.datasetID, datasetRoot))
+	return true
 }
