+++ dataset/mgmt/_rungen.go
@@ -1,0 +1,3 @@
+package gen
+
+//go:generate go run cmd/cmd.go -o types.go
