+++ dataset/mgmt/types.go
@@ -1,7 +1,7 @@
 // This file was generated by nomgen.
 // To regenerate, run `go generate` in this package.
 
-package dataset
+package mgmt
 
 import (
 	"github.com/attic-labs/noms/ref"
@@ -14,7 +14,7 @@ type DatasetSet struct {
 	s types.Set
 }
 
-type DatasetSetIterCallback (func (p Dataset) (stop bool))
+type DatasetSetIterCallback (func(p Dataset) (stop bool))
 
 func NewDatasetSet() DatasetSet {
 	return DatasetSet{types.NewSet()}
